requirements.txt
Flask==2.3.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0


torch==2.5.1
torchvision==0.20.1
Pillow==10.2.0


numpy==1.26.4
scipy==1.12.0
scikit-learn==1.3.2
joblib==1.3.2
threadpoolctl==3.2.0
annotated-types==0.7.0


typing-extensions==4.9.0
setuptools==68.2.2
wheel==0.42.0


gunicorn==21.2.0
gradio==5.46.0
